chore: rename keyboard to sendKeys
